Make example packagable and runnable with maven
