+++ examples/external/java/read-keys-example/lib/README.md
@@ -1,1 +1,1 @@
-Copy your libeelektra4j-VERSION.jar file to this directory. The same jar name should be specified in the pom.xml file [here](../pom.xml).
+Copy your libeelektra4j-VERSION.jar file to this directory.
